feat: create api for testing slack noti
